Tim Blechmann's patch: currently supercollider uses double-precision math functions for single-precision data. this leads to quite some computational overhead
this patch uses c++-style math functions, defined in cmath instead of
c89-style double-precision math functions.

the only exception is hypot, where it uses hypotf, since hypot is not
included in c++98, but is a c99 function


git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@7788 a380766d-ff14-0410-b294-a243070f3f08

+++ Source/plugins/BinaryOpUGens.cpp
@@ -604,7 +604,7 @@ void pow_d(BinaryOpUGen *unit, int inNumSamples)
 	if (inNumSamples) {
 		float a = DEMANDINPUT_A(0, inNumSamples);
 		float b = DEMANDINPUT_A(1, inNumSamples);
-		OUT0(0) = sc_isnan(a) || sc_isnan(b) ? NAN : (a < 0.f ? -pow(-a, b) : pow(a, b));
+		OUT0(0) = sc_isnan(a) || sc_isnan(b) ? NAN : (a < 0.f ? -std::pow(-a, b) : std::pow(a, b));
 	} else {
 		RESETINPUT(0);
 		RESETINPUT(1);
@@ -1022,7 +1022,7 @@ void pow_1(BinaryOpUGen *unit, int inNumSamples)
 {
 	float xa = ZIN0(0);
 	float xb = ZIN0(1);
-	ZOUT0(0) = xa >= 0.f ? pow(xa, xb) : -pow(-xa, xb);
+	ZOUT0(0) = xa >= 0.f ? std::pow(xa, xb) : -std::pow(-xa, xb);
 }
 
 
@@ -2339,7 +2339,7 @@ void pow_aa(BinaryOpUGen *unit, int inNumSamples)
 	LOOP(inNumSamples, 
 		float xa = ZXP(a);
 		float xb = ZXP(b);
-		ZXP(out) = xa >= 0.f ? pow(xa, xb) : -pow(-xa, xb);
+		ZXP(out) = xa >= 0.f ? std::pow(xa, xb) : -std::pow(-xa, xb);
 	);
 }
 
@@ -2354,13 +2354,13 @@ void pow_ak(BinaryOpUGen *unit, int inNumSamples)
 	if (xb == next_b) {
 		LOOP(inNumSamples, 
 			float xa = ZXP(a);
-			ZXP(out) = xa >= 0.f ? pow(xa, xb) : -pow(-xa, xb);
+			ZXP(out) = xa >= 0.f ? std::pow(xa, xb) : -std::pow(-xa, xb);
 		);
 	} else {
 		float slope = CALCSLOPE(next_b, xb);
 		LOOP(inNumSamples, 
 			float xa = ZXP(a);
-			ZXP(out) = xa >= 0.f ? pow(xa, xb) : -pow(-xa, xb);
+			ZXP(out) = xa >= 0.f ? std::pow(xa, xb) : -std::pow(-xa, xb);
 			xb += slope;
 		);
 		unit->mPrevB = xb;
@@ -2378,19 +2378,19 @@ void pow_ka(BinaryOpUGen *unit, int inNumSamples)
 		if (xa >= 0.f) {
 			LOOP(inNumSamples, 
 				float xb = ZXP(b);
-				ZXP(out) = pow(xa, xb);
+				ZXP(out) = std::pow(xa, xb);
 			);
 		} else {
 			LOOP(inNumSamples, 
 				float xb = ZXP(b);
-				ZXP(out) = -pow(-xa, xb);
+				ZXP(out) = -std::pow(-xa, xb);
 			);
 		}
 	} else {
 		float slope = CALCSLOPE(next_a, xa);
 		LOOP(inNumSamples, 
 			float xb = ZXP(b);
-			ZXP(out) = xa >= 0.f ? pow(xa, xb) : -pow(-xa, xb);
+			ZXP(out) = xa >= 0.f ? std::pow(xa, xb) : -std::pow(-xa, xb);
 			xa += slope;
 		);
 		unit->mPrevA = xa;
@@ -2405,7 +2405,7 @@ void pow_ia(BinaryOpUGen *unit, int inNumSamples)
 	
 	LOOP(inNumSamples, 
 		float xb = ZXP(b);
-		ZXP(out) = xa >= 0.f ? pow(xa, xb) : -pow(-xa, xb);
+		ZXP(out) = xa >= 0.f ? std::pow(xa, xb) : -std::pow(-xa, xb);
 	);
 	unit->mPrevA = xa;
 }
@@ -2419,7 +2419,7 @@ void pow_ai(BinaryOpUGen *unit, int inNumSamples)
 	
 	LOOP(inNumSamples, 
 		float xa = ZXP(a);
-		ZXP(out) = xa >= 0.f ? pow(xa, xb) : -pow(-xa, xb);
+		ZXP(out) = xa >= 0.f ? std::pow(xa, xb) : -std::pow(-xa, xb);
 	);
 	unit->mPrevB = xb;
 }
@@ -4534,7 +4534,7 @@ void atan2_aa(BinaryOpUGen *unit, int inNumSamples)
 	LOOP(inNumSamples, 
 		float xa = ZXP(a);
 		float xb = ZXP(b);
-		ZXP(out) = atan2(xa, xb);
+		ZXP(out) = std::atan2(xa, xb);
 	);
 }
 
@@ -4548,13 +4548,13 @@ void atan2_ak(BinaryOpUGen *unit, int inNumSamples)
 	if (xb == next_b) {
 		LOOP(inNumSamples, 
 			float xa = ZXP(a);
-			ZXP(out) = atan2(xa, xb);
+			ZXP(out) = std::atan2(xa, xb);
 		);
 	} else {
 		float slope = CALCSLOPE(next_b, xb);
 		LOOP(inNumSamples, 
 			float xa = ZXP(a);
-			ZXP(out) = atan2(xa, xb);
+			ZXP(out) = std::atan2(xa, xb);
 			xb += slope;
 		);
 		unit->mPrevB = xb;
@@ -4571,13 +4571,13 @@ void atan2_ka(BinaryOpUGen *unit, int inNumSamples)
 	if (xa == next_a) {
 		LOOP(inNumSamples, 
 			float xb = ZXP(b);
-			ZXP(out) = atan2(xa, xb);
+			ZXP(out) = std::atan2(xa, xb);
 		);
 	} else {
 		float slope = CALCSLOPE(next_a, xa);
 		LOOP(inNumSamples, 
 			float xb = ZXP(b);
-			ZXP(out) = atan2(xa, xb);
+			ZXP(out) = std::atan2(xa, xb);
 			xa += slope;
 		);
 		unit->mPrevA = xa;
@@ -4592,7 +4592,7 @@ void atan2_ia(BinaryOpUGen *unit, int inNumSamples)
 	
 	LOOP(inNumSamples, 
 		float xb = ZXP(b);
-		ZXP(out) = atan2(xa, xb);
+		ZXP(out) = std::atan2(xa, xb);
 	);
 	unit->mPrevA = xa;
 }
@@ -4606,7 +4606,7 @@ void atan2_ai(BinaryOpUGen *unit, int inNumSamples)
 	
 	LOOP(inNumSamples, 
 		float xa = ZXP(a);
-		ZXP(out) = atan2(xa, xb);
+		ZXP(out) = std::atan2(xa, xb);
 	);
 	unit->mPrevB = xb;
 }
@@ -4621,7 +4621,7 @@ void hypot_aa(BinaryOpUGen *unit, int inNumSamples)
 	LOOP(inNumSamples, 
 		float xa = ZXP(a);
 		float xb = ZXP(b);
-		ZXP(out) = hypot(xa, xb);
+		ZXP(out) = hypotf(xa, xb);
 	);
 }
 
@@ -4635,13 +4635,13 @@ void hypot_ak(BinaryOpUGen *unit, int inNumSamples)
 	if (xb == next_b) {
 		LOOP(inNumSamples, 
 			float xa = ZXP(a);
-			ZXP(out) = hypot(xa, xb);
+			ZXP(out) = hypotf(xa, xb);
 		);
 	} else {
 		float slope = CALCSLOPE(next_b, xb);
 		LOOP(inNumSamples, 
 			float xa = ZXP(a);
-			ZXP(out) = hypot(xa, xb);
+			ZXP(out) = hypotf(xa, xb);
 			xb += slope;
 		);
 		unit->mPrevB = xb;
@@ -4658,13 +4658,13 @@ void hypot_ka(BinaryOpUGen *unit, int inNumSamples)
 	if (xa == next_a) {
 		LOOP(inNumSamples, 
 			float xb = ZXP(b);
-			ZXP(out) = hypot(xa, xb);
+			ZXP(out) = hypotf(xa, xb);
 		);
 	} else {
 		float slope = CALCSLOPE(next_a, xa);
 		LOOP(inNumSamples, 
 			float xb = ZXP(b);
-			ZXP(out) = hypot(xa, xb);
+			ZXP(out) = hypotf(xa, xb);
 			xa += slope;
 		);
 		unit->mPrevA = xa;
@@ -4679,7 +4679,7 @@ void hypot_ia(BinaryOpUGen *unit, int inNumSamples)
 	
 	LOOP(inNumSamples, 
 		float xb = ZXP(b);
-		ZXP(out) = hypot(xa, xb);
+		ZXP(out) = hypotf(xa, xb);
 	);
 	unit->mPrevA = xa;
 }
@@ -4693,7 +4693,7 @@ void hypot_ai(BinaryOpUGen *unit, int inNumSamples)
 	
 	LOOP(inNumSamples, 
 		float xa = ZXP(a);
-		ZXP(out) = hypot(xa, xb);
+		ZXP(out) = hypotf(xa, xb);
 	);
 	unit->mPrevB = xb;
 }
